requirements.txt
fastapi>=0.68.0
pydantic>=1.8.0
uvicorn>=0.15.0
python-multipart>=0.0.5
deepface>=0.0.75
retina-face>=0.0.12
aiofiles>=0.8.0
Pillow>=9.1.1
numpy>=1.21.4